create order from selected cartitems

+++ back-end/src/main/java/com/example/eyeglass/dto/response/CodeCheckResponse.java
@@ -2,6 +2,8 @@ package com.example.eyeglass.dto.response;
 
 import lombok.Builder;
 
+import java.math.BigDecimal;
+
 @Builder
-public record CodeCheckResponse(String promoCode, int value) {
+public record CodeCheckResponse(String promoCode, BigDecimal value) {
 }
